Add Ctrl-K, Ctrl-Y from Bash
Ctrl-K "kills" (deletes) from the current input position until the
end of the line. Ctrl-Y "yanks" the last killed string, pasting it
at the current input position.

We could use the system clipboard to store the killed string, but
for consistency with Bash I store the string internally in a
global.